chore: simplified the app - less imports, less packages   * before : 1.67 Mb (466.22 Kb Gzip)   * after : 1.28 Mb (342 Kb Gzip) - sorted all imports - deleted /src/views, /src/layouts, /src/plugins - moved IndexPage.vue into App.vue - the plugins logic have been moved to main.js - if no application is running, redirect to the project's github page instead of a random about page - updated dependencies
