+++ package.json
@@ -20,25 +20,25 @@
   "dependencies": {
     "@koumoul/gh-pages-multi": "^0.7.2",
     "@mdi/font": "^7.4.47",
-    "axios": "^1.6.6",
-    "chart.js": "^4.0.0",
+    "axios": "^1.6.7",
+    "chart.js": "^4.4.1",
     "color-js": "^1.0.5",
     "debounce": "^2.0.0",
     "google-palette": "^1.1.1",
     "pinia": "^2.1.7",
     "url-polyfill": "^1.1.12",
-    "vue": "^3.3.11",
+    "vue": "^3.4.15",
     "vue-router": "^4.2.5",
     "vuetify": "^3.5.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.7",
-    "@babel/eslint-parser": "^7.23.3",
+    "@babel/core": "^7.23.9",
+    "@babel/eslint-parser": "^7.23.9",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.23.7",
-    "@babel/preset-env": "^7.23.8",
+    "@babel/plugin-transform-runtime": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
     "@data-fair/dev-server": "^1.12.0",
-    "@vitejs/plugin-vue": "^4.5.2",
+    "@vitejs/plugin-vue": "^5.0.3",
     "babel-preset-vue": "^2.0.2",
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
