--- src/plugins/moment.js
@@ -1,4 +1,0 @@
-import moment from 'moment'
-import 'moment/locale/fr'
-
-moment.locale('fr')
