--- src/plugins/vuetify.js
@@ -1,26 +1,0 @@
-import 'vuetify/styles'
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
-import colors from 'vuetify/lib/util/colors'
-import '@mdi/font/css/materialdesignicons.css'
-
-import '../styles/settings.scss'
-
-const vuetify = createVuetify({
-  components,
-  directives,
-  theme: {
-    defaultTheme: 'myTheme',
-    themes: {
-      myTheme: {
-        colors: {
-          primary: colors.blue.darken1,
-          accent: colors.orange.base
-        }
-      }
-    }
-  }
-})
-
-export default vuetify
