+++ src/router/index.js
@@ -7,12 +7,7 @@ const router = createRouter({
     {
       path: '/',
       name: 'home',
-      component: () => import('../views/IndexPage.vue')
-    },
-    {
-      path: '/about',
-      name: 'about',
-      component: () => import('../views/AboutPage.vue')
+      component: () => import('../App.vue')
     }
   ]
 })
